inv-ring: move vars loader to vars.c
