+++ src/trx/game/inventory_ring/vars.h
@@ -7,3 +7,5 @@
 extern CAMERA_INFO g_InvRing_OldCamera;
 extern INV_RING_SOURCE g_InvRing_Source[RT_NUMBER_OF];
 extern VECTOR *g_InvRing_Items;
+
+void InvRing_LoadVars(const char *path);
